redo file layout with new app flow + fix login bug with phone number
