+++ app/(auth)/(onboarding)/_layout.tsx
@@ -13,6 +13,7 @@ const OnboardingLayout = () => {
         },
       }}>
       <Stack.Screen name="index" />
+      <Stack.Screen name="signup" />
     </Stack>
   );
 };
