+++ database/queries/user.ts
@@ -1,4 +1,4 @@
-import { UserInfo } from "@/app/(auth)/(onboarding)";
+import { UserInfo } from "@/app/(auth)/(onboarding)/signup";
 import { supabase } from "@/utils/supabase";
 
 //? Getting this error when making queries with Drizzle ORM - using Supabase queries instead
@@ -7,21 +7,18 @@ import { supabase } from "@/utils/supabase";
 
 // Get user by user ID
 export const getUserByPhoneNumber = async (phoneNumber: string) => {
-  if (!phoneNumber) {
-    return { error: "Invalid input, no phoneNumber" };
-  }
-
+  // const formattedPhoneNumber = phoneNumber.slice(0); // remove the +
   try {
-    const data = await supabase
+    const { data, error } = await supabase
       .from("users")
       .select()
       .eq("phone_number", phoneNumber);
 
-    if (data.count === null) {
-      return { error: "Phone number not in database" };
+    if (error) {
+      return new Error("Error getting user by phone number");
     }
 
-    return { success: data.data[0] };
+    return data[0];
   } catch (error) {
     if (error instanceof Error) {
       console.log("Error getting user by phone number: ", error.message);
